refactor: combine all docker file into one at root
Not necessary to define docker file per client and server. They will be combined into one docker image at the very end. Also we don't need docker-compose as we are going to containerize the microservice.

Also, NGINX is not required as we can build the static client dist and put it inside the server and serve it as static content.

For consistency with other services on App Dev, I changed client port to 3000 and server port to 8080